change userserviceimpl for finding some expected exceptions

+++ d2g-service/src/main/java/com/logicify/d2g/services/UserServiceImpl.java
@@ -1,6 +1,10 @@
 package com.logicify.d2g.services;
 
 import com.logicify.d2g.domain.UserStatus;
+import com.logicify.d2g.dtos.domain.dtos.ServiceInformationDto;
+import com.logicify.d2g.dtos.domain.dtos.UserDto;
+import com.logicify.d2g.dtos.domain.exceptions.ControllerException;
+import com.logicify.d2g.dtos.domain.exceptions.ControllerExceptionCodes;
 import com.logicify.d2g.dtos.incomingdtos.UserCreateIncomingDto;
 import com.logicify.d2g.dtos.incomingdtos.UserUpdateIncomingDto;
 import com.logicify.d2g.dtos.incomingdtos.UserUpdateStatusIncomingDto;
@@ -32,55 +36,73 @@ public class UserServiceImpl implements UserService {
     private ModelMapper modelMapper;
 
     @Override
-    public String createPasswordHash(String password) throws PasswordStorage.CannotPerformOperationException {
-        return PasswordStorage.createHash(password);
+    public String createPasswordHash(String password) throws ControllerException {
+        try {
+            return PasswordStorage.createHash(password);
+        } catch (PasswordStorage.CannotPerformOperationException e) {
+            throw new ControllerException(ControllerExceptionCodes.UNCORRECTED_PASSWORD);
+        }
     }
 
     @Override
     public void createUser(UserCreateIncomingDto userCreateIncomingDto)
-            throws PasswordStorage.CannotPerformOperationException {
-        UserImpl user = modelMapper.map(userCreateIncomingDto, UserImpl.class);
-        user.setPasswordHash(createPasswordHash(userCreateIncomingDto.getPassword()));
-        user.setCreatedBy(user); //TODO: Realise getting creator from current session
-        user.setCreatedOn(ZonedDateTime.now(ZoneOffset.UTC));
-        user.setStatus(UserStatus.NEW);
-        userRepository.save(user);
+            throws ControllerException {
+        try {
+            UserImpl user = modelMapper.map(userCreateIncomingDto, UserImpl.class);
+            user.setPasswordHash(createPasswordHash(userCreateIncomingDto.getPassword()));
+            user.setCreatedBy(user); //TODO: Realise getting creator from current session
+            user.setCreatedOn(ZonedDateTime.now(ZoneOffset.UTC));
+            user.setStatus(UserStatus.NEW);
+            userRepository.save(user);
+        } catch (ControllerException e) {
+            throw new ControllerException(ControllerExceptionCodes.UNCORRECTED_PASSWORD);
+        }
     }
 
     @Override
     public UsersListOutgoingDto findAll() {
-        return getListUserFromRepository(userRepository.findAll());
+        UsersListOutgoingDto usersListOutgoingDto = getListUserFromRepository(userRepository.findAll());
+        usersListOutgoingDto.setService(new ServiceInformationDto());
+        return usersListOutgoingDto;
     }
 
     @Override
-    public UserOutgoingDto findOne(UUID id) {
-        //if (!userRepository.exists(id)) throw new Exception(); TODO:Throwing exception if user not exist
+    public UserOutgoingDto findOne(UUID id) throws ControllerException {
+        if (!userRepository.exists(id)) throw new ControllerException(ControllerExceptionCodes.USER_NOT_EXIST);
         UserImpl user = userRepository.findOne(id);
-        return modelMapper.map(user, UserOutgoingDto.class);
+        UserDto userDto = modelMapper.map(user, UserDto.class);
+        UserOutgoingDto userOutgoingDto = new UserOutgoingDto();
+        userOutgoingDto.setUserDto(userDto);
+        userOutgoingDto.setService(new ServiceInformationDto());
+        return userOutgoingDto;
     }
 
     @Override
-    public void delete(UUID id) {
+    public void delete(UUID id) throws ControllerException {
+        if (!userRepository.exists(id)) throw new ControllerException(ControllerExceptionCodes.USER_NOT_EXIST);
         userRepository.delete(id); //TODO: Throwing exception if user not exist
     }
 
     @Override
-    public void updateUser(UUID id, UserUpdateIncomingDto userUpdateIncomingDto)
-            throws PasswordStorage.CannotPerformOperationException {
-        UserImpl user = userRepository.findOne(id);
-        if (userUpdateIncomingDto.getFirstName() != null)
-            user.setFirstName(userUpdateIncomingDto.getFirstName());
-        if (userUpdateIncomingDto.getLastName() != null)
-            user.setLastName(userUpdateIncomingDto.getLastName());
-        if (userUpdateIncomingDto.getAvatarUrl() != null)
-            user.setAvatarUrl(userUpdateIncomingDto.getAvatarUrl());
-        if (userUpdateIncomingDto.getEmail() != null)
-            user.setEmail(userUpdateIncomingDto.getEmail());
-        if (userUpdateIncomingDto.getPassword() != null)
-            user.setPasswordHash(createPasswordHash(userUpdateIncomingDto.getPassword()));
-        user.setUpdatedOn(ZonedDateTime.now(ZoneOffset.UTC));
-        user.setUpdatedBy(user); //TODO: Realise getting updater from current session
-        userRepository.save(user);
+    public void updateUser(UUID id, UserUpdateIncomingDto userUpdateIncomingDto) throws ControllerException {
+        try {
+            UserImpl user = userRepository.findOne(id);
+            if (userUpdateIncomingDto.getFirstName() != null)
+                user.setFirstName(userUpdateIncomingDto.getFirstName());
+            if (userUpdateIncomingDto.getLastName() != null)
+                user.setLastName(userUpdateIncomingDto.getLastName());
+            if (userUpdateIncomingDto.getAvatarUrl() != null)
+                user.setAvatarUrl(userUpdateIncomingDto.getAvatarUrl());
+            if (userUpdateIncomingDto.getEmail() != null)
+                user.setEmail(userUpdateIncomingDto.getEmail());
+            if (userUpdateIncomingDto.getPassword() != null)
+                user.setPasswordHash(createPasswordHash(userUpdateIncomingDto.getPassword()));
+            user.setUpdatedOn(ZonedDateTime.now(ZoneOffset.UTC));
+            user.setUpdatedBy(user); //TODO: Realise getting updater from current session
+            userRepository.save(user);
+        } catch (ControllerException e) {
+            throw new ControllerException(ControllerExceptionCodes.UNCORRECTED_PASSWORD);
+        }
     }
 
     @Override
@@ -95,10 +117,11 @@ public class UserServiceImpl implements UserService {
     private UsersListOutgoingDto getListUserFromRepository(Iterable<UserImpl> userIterable) {
         List<UserImpl> users = new ArrayList<>();
         userIterable.forEach(users::add);
-        List<UserOutgoingDto> outgoingDtos = new ArrayList<>();
-        users.forEach(user -> outgoingDtos.add(modelMapper.map(user, UserOutgoingDto.class)));
+        List<UserDto> outgoingDtos = new ArrayList<>();
+        users.forEach(user -> outgoingDtos.add(modelMapper.map(user, UserDto.class)));
         UsersListOutgoingDto usersListOutgoingDto = new UsersListOutgoingDto();
-        usersListOutgoingDto.setUserOutgoingDtoList(outgoingDtos);
+        usersListOutgoingDto.setUserDtoList(outgoingDtos);
+        usersListOutgoingDto.setService(new ServiceInformationDto());
         return usersListOutgoingDto;
     }
